move publish_datasets function away from utils.py
move beginning and ending logic from publish_datasets to publish_datasets_v2
TODO: need to rename publish_datasets_v2 to something else
